SCS-146 Done

+++ Aeroclub.Cargo/Aeroclub.Cargo.Core/Entities/Airport.cs
@@ -7,8 +7,8 @@ namespace Aeroclub.Cargo.Core.Entities
     {
         public string Name { get; set; }  = null!;
         public string Code { get; set; }  = null!;
-        public decimal Lat { get; set; }
-        public decimal Lon { get; set; }
+        public double Lat { get; set; }
+        public double Lon { get; set; }
         public Guid CountryId { get; set; }
 
         public virtual Country Country { get; set; }
